+++ Dockerfile
@@ -12,7 +12,7 @@ ENV MYSQL_DB=${MYSQL_DB} \
     MYSQL_PORT=${MYSQL_PORT} \
     MYSQL_USER=${MYSQL_USER} \
     MYSQL_PASSWORD=${MYSQL_PASSWORD} \
-    PYTHONPATH=${PYTHONPATH}:/app/itou_c4_api
+    PYTHONPATH=${PYTHONPATH}:/app/lemarche
 
 WORKDIR /app
 
@@ -47,10 +47,10 @@ RUN poetry config virtualenvs.create false && \
     poetry install $(test $ENV == "prod" && echo "--no-dev") --no-interaction --no-ansi
 
 # CMD ["bash"]
-CMD ["itou_c4_api/runner.sh"]
+CMD ["lemarche/runner.sh"]
 
 # # Multistage build : RUN
 # FROM base as final
 # 
 # # CMD ["bash"]
-# CMD ["itou_c4_api/runner.sh"]
+# CMD ["lemarche/runner.sh"]
