+++ lemarche/api/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class UsersConfig(AppConfig):
+class C4DirectoryConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "itou_c4_api.users"
+    name = "lemarche.api"
