+++ lemarche/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for itou_c4_api project.
+ASGI config for lemarche project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -12,6 +12,6 @@ import os
 from django.core.asgi import get_asgi_application
 
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "itou_c4_api.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "lemarche.settings")
 
 application = get_asgi_application()
