+++ lemarche/cocorico/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class CocoricoConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "itou_c4_api.cocorico"
+    name = "lemarche.cocorico"
