+++ lemarche/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for itou_c4_api project.
+Django settings for lemarche project.
 
 Generated by 'django-admin startproject' using Django 3.2.1.
 
@@ -49,11 +49,11 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     "rest_framework",
     "drf_spectacular",
-    "itou_c4_api",
-    "itou_c4_api.c4_directory",
-    "itou_c4_api.users",
-    "itou_c4_api.cocorico",
-    # "c4_directory.apps.C4DirectoryConfig",
+    "lemarche",
+    "lemarche.api",
+    "lemarche.users",
+    "lemarche.cocorico",
+    # "api.apps.C4DirectoryConfig",
     # "users.apps.UsersConfig",
     # "cocorico.apps.CocoricoConfig",
 ]
@@ -68,7 +68,7 @@ MIDDLEWARE = [
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 ]
 
-ROOT_URLCONF = "itou_c4_api.urls"
+ROOT_URLCONF = "lemarche.urls"
 
 TEMPLATES = [
     {
@@ -87,7 +87,7 @@ TEMPLATES = [
 ]
 
 
-WSGI_APPLICATION = "itou_c4_api.wsgi.application"
+WSGI_APPLICATION = "lemarche.wsgi.application"
 
 
 # Database
@@ -115,7 +115,7 @@ DATABASES = {
 }
 
 # Needed as long as Cocorico database used as data source
-DATABASE_ROUTERS = ["itou_c4_api.routers.CocoRouter"]
+DATABASE_ROUTERS = ["lemarche.routers.CocoRouter"]
 
 
 # Password validation
