+++ lemarche/users/admin.py
@@ -1,8 +1,8 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
 from django.contrib.auth.models import Group
-from itou_c4_api.users.forms import UserChangeForm, UserCreationForm
-from itou_c4_api.users.models import User
+from lemarche.users.forms import UserChangeForm, UserCreationForm
+from lemarche.users.models import User
 
 
 class UserAdmin(UserAdmin):
