+++ lemarche/users/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class C4DirectoryConfig(AppConfig):
+class UsersConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "itou_c4_api.c4_directory"
+    name = "lemarche.users"
