+++ lemarche/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for itou_c4_api project.
+WSGI config for lemarche project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -12,6 +12,6 @@ import os
 from django.core.wsgi import get_wsgi_application
 
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "itou_c4_api.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "lemarche.settings")
 
 application = get_wsgi_application()
